// types/types.ts
export type PostProps = {
    title?: string;
    slug?: string;
    author?: string;
    description?: any;
    mainImage?: any;
    price?: any;
    body?: any;
    dimensions?: string;
    isReversed?: boolean;
};

export type ProductsProps = {
    title?: string;
    slug?: string;
    description?: any;
    mainImage?: any;
    price?: any;
    dimensions?: string;
};